Move class to mock package
