+++ libraries/test_utils/src/main/java/com/cryart/sabbathschool/test/di/mock/MockDataFactory.kt
@@ -1,4 +1,26 @@
-package com.cryart.sabbathschool.test.di.repository
+/*
+ * Copyright (c) 2021. Adventech <info@adventech.io>
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+
+package com.cryart.sabbathschool.test.di.mock
 
 import app.ss.lessons.data.model.SSBibleVerses
 import app.ss.lessons.data.model.SSDay
@@ -13,8 +35,6 @@ import app.ss.lessons.data.model.WeekDay
  * Mock data for UI tests
  */
 object MockDataFactory {
-
-    const val QUARTERLY_INDEX = ""
 
     val versions = listOf(
         "NASB", "NKJV", "KJV"
